pyproject.toml
[project]
name = "mcp-server-box"
version = "0.7.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
scripts = { "mcp-server-box" = "mcp_server_box:main" }
license = "MIT"
authors = [{ name = "Box DevRel", email = "dx@box.com" }]
dependencies = [
    "box-ai-agents-toolkit>=0.1.5",
    "colorlog>=6.10.1",
    "fastapi>=0.121.0",
    "mcp[cli]>=1.19.0",
    "python-pptx>=1.0.2",
    "python-dotenv>=1.2.1",
    "tomli>=2.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.3",
]

[tool.pytest.ini_options]
pythonpath = [".", "src", "tests"]
